# swagger.json
{"openapi":"3.0.0","info":{"title":"Movie Reviewer","version":"0.1.0"},"paths":{"/api/movies":{"get":{"description":"Get all movies","parameters":[{"in":"query","name":"query","required":false,"description":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"A successful response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Movie"}}}}}},"tags":["movies"],"produces":["application/json"]}},"/api/movies/{id}":{"get":{"description":"Get all movies","parameters":[{"in":"path","name":"id","required":true,"description":"Movie id","schema":{"type":"string"}}],"responses":{"200":{"description":"A successful response","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/Movie"}}}}},"tags":["movies"],"produces":["application/json"]}},"/api/movies/{id}/comment":{"post":{"description":"Get all movies","parameters":[{"in":"path","name":"id","required":true,"description":"Movie id","schema":{"type":"string"}}],"requestBody":{"description":"Comment content","required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCommentDto"}}}},"responses":{"200":{"description":"A successful response","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/Comment"}}}}},"tags":["movies"],"produces":["application/json"]}},"/api/movies/{id}/rate":{"post":{"description":"Rate a movie","parameters":[{"in":"path","name":"id","required":true,"description":"Movie id","schema":{"type":"string"}}],"requestBody":{"description":"Rate content","required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RateMovieDto"}}}},"responses":{"200":{"description":"A successful response","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/Movie"}}}}},"tags":["movies"],"produces":["application/json"]}},"/api/users/search":{"get":{"description":"Get user by id","parameters":[{"in":"query","name":"query","required":true,"description":"query to search","schema":{"type":"string"}}],"responses":{"200":{"description":"A successful response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}},"tags":["users"],"produces":["application/json"]}},"/api/users/{id}":{"get":{"description":"Get user by id","parameters":[{"in":"path","name":"id","required":true,"description":"user id","schema":{"type":"string"}}],"responses":{"200":{"description":"A successful response","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/User"}}}}},"tags":["users"],"produces":["application/json"]}},"/api/users/signup":{"post":{"description":"Signup user","requestBody":{"description":"Create user dto","required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"A successful response","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}}}}}}},"tags":["users"],"produces":["application/json"]}},"/api/users/login":{"post":{"description":"login user","requestBody":{"description":"Create user dto","required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string","format":"email"},"password":{"type":"string","format":"password"}}}}}},"responses":{"200":{"description":"A successful response","content":{"application/json":{"schema":{"type":"object","properties":{"token":{"type":"string"}}}}}}},"tags":["users"],"produces":["application/json"]}},"/api/users/favorites":{"post":{"description":"Add movie to favorites","requestBody":{"description":"Add to favorites dto","required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddToFavoritesDto"}}}},"responses":{"200":{"description":"A successful response","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}}}}}}},"tags":["users"],"produces":["application/json"]}},"/api/users/favorites/{id}":{"delete":{"description":"Delete movie from favorites","parameters":[{"in":"path","name":"id","required":true,"description":"favorite movie id","schema":{"type":"string"}}],"responses":{"200":{"description":"A successful response","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}}}}}}},"tags":["users"],"produces":["application/json"]}},"/api/users/followers":{"post":{"description":"Add user to followers","requestBody":{"description":"Add to followers dto","required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FollowDto"}}}},"responses":{"201":{"description":"A successful response","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}}}}}}},"tags":["users"],"produces":["application/json"]}},"/api/users":{"patch":{"description":"Modify user","requestBody":{"description":"Modify user dto","required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModifyUserDto"}}}},"responses":{"200":{"description":"A successful response","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}}}}}}},"tags":["users"],"produces":["application/json"]}},"/api/users/upload-profile":{"post":{"description":"Upload profile picture","requestBody":{"description":"Upload profile picture dto","required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"picture":{"type":"string","format":"binary","description":"Profile picture"}}}}}},"responses":{"200":{"description":"A successful response","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}}}}}}},"tags":["users"],"produces":["application/json"]}}},"components":{"schemas":{"CreateCommentDto":{"type":"object","properties":{"comment":{"type":"string","description":"Comment content","maxLength":255,"minLength":1}}},"RateMovieDto":{"type":"object","properties":{"score":{"type":"integer","description":"Score of the movie","format":"int32","minimum":1,"maximum":5}}},"AddToFavoritesDto":{"type":"object","properties":{"movieId":{"type":"string","description":"Movie id","format":"ObjectId"}}},"CreateUserDto":{"type":"object","properties":{"email":{"type":"string","description":"User email","format":"email"},"name":{"type":"string","description":"User name"},"dateOfBirth":{"type":"string","description":"User date of birth","format":"date-time"},"password":{"type":"string","description":"User password","minLength":8}}},"FollowDto":{"type":"object","properties":{"userId":{"type":"string","description":"User id to follow","format":"ObjectId"}}},"ModifyUserDto":{"type":"object","properties":{"name":{"type":"string","description":"User name","required":true}}},"Comment":{"type":"object","properties":{"_id":{"type":"string","description":"Comment id"},"belongsTo":{"type":"string","description":"Movie id"},"createdBy":{"$ref":"#/components/schemas/User"},"content":{"type":"string","description":"Comment content"}}},"Platform":{"type":"object","properties":{"name":{"type":"string","description":"Platform of the movie"},"url":{"type":"string","description":"Url of the movie","format":"url"}}},"Movie":{"type":"object","properties":{"_id":{"type":"string"},"name":{"type":"string","description":"Movie title"},"cover":{"type":"string","description":"Movie cover","format":"url"},"synopsis":{"type":"string","description":"Movie synopsis"},"trailer":{"type":"string","description":"Movie trailer","format":"url"},"score":{"type":"number","description":"Movie score"},"releaseDate":{"type":"string","description":"Movie release date","format":"date-time"},"platforms":{"type":"array","items":{"$ref":"#/components/schemas/Platform"},"description":"Movie platforms"},"screenwriters":{"type":"array","items":{"type":"string"},"description":"Movie screen writers"},"cast":{"type":"array","items":{"type":"string"},"description":"Movie cast"},"genres":{"type":"array","items":{"type":"string"},"description":"Movie genres"},"comments":{"type":"array","items":{"$ref":"#/components/schemas/Comment"},"description":"Movie comments"}}},"User":{"type":"object","properties":{"email":{"type":"string","description":"User email","format":"email"},"name":{"type":"string","description":"User name"},"dateOfBirth":{"type":"string","description":"User date of birth","format":"date-time"},"avatar":{"type":"string","description":"User avatar","format":"url"},"following":{"type":"array","items":{"$ref":"#/components/schemas/User"}},"favorites":{"type":"array","items":{"$ref":"#/components/schemas/Movie"}}}}}},"tags":[]}